move abi.util classes to util
