+++ src/main/java/com/esaulpaugh/headlong/abi/ABIJSON.java
@@ -32,10 +32,10 @@ import java.util.EnumSet;
 import java.util.List;
 import java.util.Set;
 
-import static com.esaulpaugh.headlong.abi.util.JsonUtils.getArray;
-import static com.esaulpaugh.headlong.abi.util.JsonUtils.getBoolean;
-import static com.esaulpaugh.headlong.abi.util.JsonUtils.getString;
-import static com.esaulpaugh.headlong.abi.util.JsonUtils.parseArray;
+import static com.esaulpaugh.headlong.util.JsonUtils.getArray;
+import static com.esaulpaugh.headlong.util.JsonUtils.getBoolean;
+import static com.esaulpaugh.headlong.util.JsonUtils.getString;
+import static com.esaulpaugh.headlong.util.JsonUtils.parseArray;
 
 /** For parsing JSON representations of {@link ABIObject}s according to the ABI specification. */
 public final class ABIJSON {
