+++ src/main/java/com/esaulpaugh/headlong/abi/ContractError.java
@@ -15,7 +15,7 @@
 */
 package com.esaulpaugh.headlong.abi;
 
-import com.esaulpaugh.headlong.abi.util.JsonUtils;
+import com.esaulpaugh.headlong.util.JsonUtils;
 import com.google.gson.JsonObject;
 
 import java.util.Objects;
