+++ src/main/java/com/esaulpaugh/headlong/util/JsonUtils.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package com.esaulpaugh.headlong.abi.util;
+package com.esaulpaugh.headlong.util;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
