+++ src/test/java/com/esaulpaugh/headlong/abi/AdvancedABICasesTest.java
@@ -13,9 +13,9 @@
 package com.esaulpaugh.headlong.abi;
 
 import com.esaulpaugh.headlong.TestUtils;
-import com.esaulpaugh.headlong.abi.util.Deserializer;
+import com.esaulpaugh.headlong.util.Deserializer;
 import com.esaulpaugh.headlong.util.FastHex;
-import com.esaulpaugh.headlong.abi.util.JsonUtils;
+import com.esaulpaugh.headlong.util.JsonUtils;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
