+++ src/test/java/com/esaulpaugh/headlong/abi/EncodeTest.java
@@ -16,8 +16,8 @@
 package com.esaulpaugh.headlong.abi;
 
 import com.esaulpaugh.headlong.TestUtils;
-import com.esaulpaugh.headlong.abi.util.Uint;
-import com.esaulpaugh.headlong.abi.util.WrappedKeccak;
+import com.esaulpaugh.headlong.util.Uint;
+import com.esaulpaugh.headlong.util.WrappedKeccak;
 import com.esaulpaugh.headlong.util.Strings;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Disabled;
