+++ src/test/java/com/esaulpaugh/headlong/abi/FunctionTest.java
@@ -16,7 +16,7 @@
 package com.esaulpaugh.headlong.abi;
 
 import com.esaulpaugh.headlong.TestUtils;
-import com.esaulpaugh.headlong.abi.util.WrappedKeccak;
+import com.esaulpaugh.headlong.util.WrappedKeccak;
 import com.esaulpaugh.headlong.util.Strings;
 import org.junit.jupiter.api.Test;
