+++ src/test/java/com/esaulpaugh/headlong/abi/PackedEncoderTest.java
@@ -16,8 +16,8 @@
 package com.esaulpaugh.headlong.abi;
 
 import com.esaulpaugh.headlong.TestUtils;
-import com.esaulpaugh.headlong.abi.util.BizarroInts;
-import com.esaulpaugh.headlong.abi.util.Uint;
+import com.esaulpaugh.headlong.util.BizarroInts;
+import com.esaulpaugh.headlong.util.Uint;
 import com.esaulpaugh.headlong.util.FastHex;
 import com.esaulpaugh.headlong.util.Strings;
 import org.junit.jupiter.api.Test;
